refactor: extract AgentWorkspace props
